requirement.txt
python-telegram-bot
http://effbot.org/media/downloads/xmlrpclib-1.0.1.zip